Update GraphQl functions with new SemaphoreSlim model.
The GraphQl functions were not using SemaphoreSlim properly.  All three methods were condensed into one method for the user of the client.
GetGraphQlInChunksAsync was not doing anything different than GetGraphQlAsyncInParallel really so the original functionality of GetGraphQlInChunksAsync was removed.  Now GetGraphQlInChunksAsync is a method that calls both GetGraphQlAsyncInParallel and ProcessSemaphoreSlimsForGraphQl for the user so they don't have to call two separate methods.  GetGraphQlAsyncInParallel and ProcessSemaphoreSlimsForGraphQl are now private methods.